Fix the structure ...
