+++ src/slickml/__init__.py
@@ -1,8 +1,10 @@
 """
+
 SlickML
 =======
 
 """
+# TODO(amir): figure out a good way to pull-off the doc here
 from importlib.metadata import version
 
 __version__ = version("slickml")
