+++ src/slickml/classification/_todo.py
@@ -5,8 +5,9 @@ import shap
 import xgboost as xgb
 from sklearn.preprocessing import StandardScaler
 
-from slickml.formatting import Color
-from slickml.plotting import (
+from slickml.utils import df_to_csr
+from slickml.utils._todo_fmt import Color
+from slickml.visualization._todo_plotting import (
     plot_glmnet_coeff_path,
     plot_glmnet_cv_results,
     plot_shap_summary,
@@ -14,12 +15,11 @@ from slickml.plotting import (
     plot_xgb_cv_results,
     plot_xgb_feature_importance,
 )
-from slickml.utils import df_to_csr
 
 
-class XGBoostRegressor:
-    """XGBoost Regressor.
-    This is wrapper using XGBoost regressor to train a XGBoost
+class XGBoostClassifier:
+    """XGBoost Classifier.
+    This is wrapper using XGBoost classifier to train a XGBoost
     model with using number of boosting rounds from the inputs. This
     function is pretty useful when feature selection is done and you
     want to train a model on the whole data and test on a separate
@@ -31,11 +31,11 @@ class XGBoostRegressor:
     num_boost_round: int, optional (default=200)
         Number of boosting round to train the model
 
-    metrics: str or tuple[str], optional (default=("rmse"))
+    metrics: str or tuple[str], optional (default=("auc"))
         Metric used for evaluation at cross-validation
         using xgboost.cv(). Please note that this is different
         than eval_metric that needs to be passed to params dict.
-        Possible values are "rmse", "rmsle", "mae"
+        Possible values are "auc", "aucpr", "error", "logloss"
 
     sparse_matrix: bool, optional (default=False)
         Flag to convert data to sparse matrix with csr format.
@@ -58,9 +58,9 @@ class XGBoostRegressor:
 
     params: dict, optional
         Set of parameters for evaluation of xboost.train()
-        (default={"eval_metric" : "rmse",
+        (default={"eval_metric" : "auc",
                   "tree_method": "hist",
-                  "objective" : "reg:squarederror",
+                  "objective" : "binary:logistic",
                   "learning_rate" : 0.05,
                   "max_depth": 2,
                   "min_child_weight" : 1,
@@ -70,8 +70,8 @@ class XGBoostRegressor:
                   "subsample" : 0.9,
                   "max_delta_step": 1,
                   "verbosity" : 0,
-                  "nthread" : 4})
-        Other options for objective: "reg:logistic", "reg:squaredlogerror"
+                  "nthread" : 4,
+                  "scale_pos_weight" : 1})
 
     Attributes
     ----------
@@ -108,8 +108,16 @@ class XGBoostRegressor:
         Returns None and applies the training process using
         the (X_train, y_train) set using xgboost.train()
 
-    predict(X_test, y_test): instance method
-        Return the predicted target values.
+    predict_proba(X_test, y_test): instance method
+        Return the prediction probabilities for positive classes. Please note that
+        it only reports the probability of the positive class, while the sklearn
+        one returns for both and slicing like pred_proba[:, 1]
+        is needed for positive class predictions
+
+    predict(X_test, y_test, threshold=0.5): instance method
+        Return the prediction classes based on the passed threshold.
+        The default threshold is set at 0.5 while you can find the optimum thresholds
+        based on different methods using BinaryClassificationMetrics.
 
     get_params(): instance method
         Returns params dict
@@ -144,14 +152,12 @@ class XGBoostRegressor:
                 self.num_boost_round = num_boost_round
 
         if metrics is None:
-            self.metrics = "rmse"
+            self.metrics = "auc"
         else:
             if not isinstance(metrics, str):
-                raise TypeError("The input metric must be a str dtype.")
+                raise TypeError("The input metrics must be a str dtype.")
             else:
-                # TODO: update metric in next API update
-                # mape, mphe should be added
-                if metrics in ["rmse", "rmsle", "mae"]:
+                if metrics in ["auc", "aucpr", "error", "logloss"]:
                     self.metrics = metrics
                 else:
                     raise ValueError("The input metric value is not valid.")
@@ -188,9 +194,9 @@ class XGBoostRegressor:
                 else:
                     raise ValueError("The input importance_type value is not valid.")
         params_ = {
-            "eval_metric": "rmse",
+            "eval_metric": "auc",
             "tree_method": "hist",
-            "objective": "reg:squarederror",
+            "objective": "binary:logistic",
             "learning_rate": 0.05,
             "max_depth": 2,
             "min_child_weight": 1,
@@ -201,6 +207,7 @@ class XGBoostRegressor:
             "max_delta_step": 1,
             "verbosity": 0,
             "nthread": 4,
+            "scale_pos_weight": 1,
         }
         if params is None:
             self.params = params_
@@ -252,9 +259,13 @@ class XGBoostRegressor:
 
         return self.feature_importance_
 
-    def predict(self, X_test, y_test=None):
+    def predict_proba(self, X_test, y_test=None):
         """
-        Function to return the prediction of target values.
+        Function to return the prediction probabilities for both classes.
+        Please note that it only reports the probability of the positive class,
+        while the sklearn one returns for both and slicing like pred_proba[:, 1]
+        is needed for positive class predictions. Note that y_test is optional while
+        it might not be available in validiation.
 
         Parameters
         ----------
@@ -265,7 +276,31 @@ class XGBoostRegressor:
             List of validation ground truth binary values [0, 1]
         """
         self.dtest_ = self._dtest(X_test, y_test)
-        self.y_pred_ = self.model_.predict(self.dtest_, output_margin=False)
+        self.y_pred_proba_ = self.model_.predict(self.dtest_, output_margin=False)
+
+        return self.y_pred_proba_
+
+    def predict(self, X_test, y_test=None, threshold=0.5):
+        """
+        Function to return the prediction classes based on the passed threshold.
+        The default threshold is set at 0.5 while you can find the optimum thresholds
+        based on different methods using BinaryClassificationMetrics.
+
+        Parameters
+        ----------
+        X_test: numpy.array or pandas.DataFrame
+            Validation features data
+
+        y_test: numpy.array[int] or list[int], optional (default=None)
+            List of validation ground truth binary values [0, 1]
+
+        threshold: float, optional (default=0.5)
+            Threshold to define classes based on probabilities.
+            predict_proba >= threshold would be defined as 1 else 0.
+        """
+        self.dtest_ = self._dtest(X_test, y_test)
+        self.y_pred_proba_ = self.model_.predict(self.dtest_, output_margin=False)
+        self.y_pred_ = [1 if p >= threshold else 0 for p in self.y_pred_proba_]
 
         return self.y_pred_
 
@@ -284,7 +319,7 @@ class XGBoostRegressor:
 
         """Function to plot XGBoost feature importance.
         This function is a helper function based on the feature_importance_
-        attribute of the XGBoostRegressor class.
+        attribute of the XGBoostCVClassifier class.
 
         Parameters
         ----------
@@ -694,9 +729,9 @@ class XGBoostRegressor:
         return df
 
 
-class XGBoostCVRegressor(XGBoostRegressor):
-    """XGBoost CV Regressor.
-    This is subclass of XGBoostRegressor to run xgboost.cv()
+class XGBoostCVClassifier(XGBoostClassifier):
+    """XGBoost CV Classifier.
+    This is subclass of XGBoostClassifier to run xgboost.cv()
     model with n-folds cross-validation and train model based on
     the best number of boosting round to avoid over-fitting. This
     function is pretty useful when feature selection is done and you
@@ -714,11 +749,11 @@ class XGBoostCVRegressor(XGBoostRegressor):
     n_splits: int, optional (default=4)
         Number of folds for cross-validation
 
-    metrics: str or tuple[str], optional (default=("rmse"))
+    metrics: str or tuple[str], optional (default=("auc"))
         Metric used for evaluation at cross-validation
         using xgboost.cv(). Please note that this is different
         than eval_metric that needs to be passed to params dict.
-        Possible values are "rmse", "rmse", "mae"
+        Possible values are "auc", "aucpr", "error", "logloss"
 
     early_stopping_rounds: int, optional (default=20)
         The criterion to early abort the xgboost.cv() phase
@@ -726,6 +761,11 @@ class XGBoostCVRegressor(XGBoostRegressor):
 
     random_state: int, optional (default=1367)
         Random seed
+
+    stratified: bool, optional (default=True)
+        Flag to stratificaiton of the targets to run xgboost.cv() to
+        find the best number of boosting round at each fold of
+        each iteration
 
     shuffle: bool, optional (default=True)
         Flag to shuffle data to have the ability of building
@@ -752,9 +792,9 @@ class XGBoostCVRegressor(XGBoostRegressor):
 
     params: dict, optional
         Set of parameters for evaluation of xboost.train()
-        (default={"eval_metric" : "rmse",
+        (default={"eval_metric" : "auc",
                   "tree_method": "hist",
-                  "objective" : "reg:squarederror",
+                  "objective" : "binary:logistic",
                   "learning_rate" : 0.05,
                   "max_depth": 2,
                   "min_child_weight" : 1,
@@ -764,8 +804,8 @@ class XGBoostCVRegressor(XGBoostRegressor):
                   "subsample" : 0.9,
                   "max_delta_step": 1,
                   "verbosity" : 0,
-                  "nthread" : 4})
-        Other options for objective: "reg:logistic", "reg:squaredlogerror"
+                  "nthread" : 4,
+                  "scale_pos_weight" : 1})
 
     callbacks: bool, optional (default=False)
         Flag for printing results during xgboost.cv().
@@ -817,8 +857,16 @@ class XGBoostCVRegressor(XGBoostRegressor):
         Returns None and applies the training process using
         the (X_train, y_train) set using xgboost.cv() and xgboost.train()
 
-    predict(X_test, y_test): instance method
-        Return the predicted target values.
+    predict_proba(X_test, y_test): instance method
+        Return the prediction probabilities for positive classes. Please note that
+        it only reports the probability of the positive class, while the sklearn
+        one returns for both and slicing like pred_proba[:, 1]
+        is needed for positive class predictions
+
+    predict(X_test, y_test, threshold=0.5): instance method
+        Return the prediction classes based on the passed threshold.
+        The default threshold is set at 0.5 while you can find the optimum thresholds
+        based on different methods using BinaryClassificationMetrics.
 
     get_params(): instance method
         Returns params dict
@@ -848,6 +896,7 @@ class XGBoostCVRegressor(XGBoostRegressor):
         metrics=None,
         early_stopping_rounds=None,
         random_state=None,
+        stratified=True,
         shuffle=True,
         sparse_matrix=False,
         scale_mean=False,
@@ -890,6 +939,11 @@ class XGBoostCVRegressor(XGBoostRegressor):
                 raise TypeError("The input random_state must have integer dtype.")
             else:
                 self.random_state = random_state
+
+        if not isinstance(stratified, bool):
+            raise TypeError("The input stratified must have bool dtype.")
+        else:
+            self.stratified = stratified
 
         if not isinstance(shuffle, bool):
             raise TypeError("The input shuffle must have bool dtype.")
@@ -1050,6 +1104,7 @@ class XGBoostCVRegressor(XGBoostRegressor):
             dtrain=self.dtrain_,
             num_boost_round=self.num_boost_round,
             nfold=self.n_splits,
+            stratified=self.stratified,
             metrics=self.metrics,
             early_stopping_rounds=self.early_stopping_rounds,
             seed=self.random_state,
@@ -1072,12 +1127,13 @@ class XGBoostCVRegressor(XGBoostRegressor):
         return model
 
 
-class GLMNetCVRegressor:
-    """GLMNet CV Regressor.
+class GLMNetCVClassifier:
+    """GLMNet CV Classifier.
     This is wrapper using GLM-Net to train a Regularized Linear Model
-    and find the optimal penalty values through N-Folds cross validation.
-    This function is pretty useful to train a Elastic-Net model with
-    the ability of feature reduction. Main theoretical reference:
+    via logitic regression and find the optimal penalty values through
+    N-Folds cross validation. This function is pretty useful to train
+    a Logit-Net model with the ability of feature reduction. Main
+    theoretical reference:
     (https://web.stanford.edu/~hastie/glmnet/glmnet_alpha.html)
 
     Parameters
@@ -1093,10 +1149,10 @@ class GLMNetCVRegressor:
         determining lambda_best_ and lambda_max_. If non-zero, must be
         at least 3.
 
-    metric: str or callable, optional (default="r2")
+    metric: str or callable, optional (default="auc")
         Metric used for model selection during cross validation.
-        Valid options are "r2", "mean_squared_error", "mean_absolute_error",
-        and "median_absolute_error". Alternatively, supply a function or callable
+        Valid options are "accuracy", "roc_auc" (alias = "auc"), "average_precision",
+        "precision", "recall". Alternatively, supply a function or callable
         object with the following signature "scorer(estimator, X, y)".
         Note, the metric function affects the selection of "lambda_best_" and
         "lambda_max_", fitting the same data ith different metric methods will result
@@ -1154,11 +1210,11 @@ class GLMNetCVRegressor:
     X_test_: pandas.DataFrame()
         Returns transformed testing data set.
 
-    y_train_: numpy.array[float] or list[float]
-        Returns the list of training ground truth target values
+    y_train_: numpy.array[int] or list[int]
+        Returns the list of training ground truth binary values [0, 1]
 
     y_test_: numpy.array[int] or list[int]
-        Returns the list of testing ground truth target values
+        Returns the list of testing ground truth binary values [0, 1]
 
     coeff_: pandas.DataFrame
         Return the model's non-zero coefficients
@@ -1177,10 +1233,17 @@ class GLMNetCVRegressor:
 
     fit(X_train, y_train): instance method
         Returns None and applies the training process using
-        the (X_train, y_train) set using glmnet.ElasticNet()
+        the (X_train, y_train) set using glmnet.LogitNet()
 
-    predict(X_test, y_test=None): instance method
-        Returns the prediction values
+    predict_proba(X_test, y_test=None): instance method
+        Return the prediction probabilities for positive classes. Please note that
+        it only reports the probability of the positive class, while the sklearn
+        one returns for both and slicing like pred_proba[:, 1]
+        is needed for positive class predictions
+
+    predict(X_test, y_test=None, threshold=0.5): instance method
+        Returns the prediction classes based on the passed threshold.
+        The default threshold is set at 0.5.
 
     get_params(): instance method
         Returns params dict
@@ -1247,18 +1310,21 @@ class GLMNetCVRegressor:
                 self.n_splits = n_splits
 
         if metric is None:
-            self.metric = "r2"
+            self.metric = "roc_auc"
         else:
             if not isinstance(metric, str):
                 raise TypeError("The input metric must have str dtype.")
             else:
                 if metric in [
-                    "r2",
-                    "mean_squared_error",
-                    "mean_absolute_error",
-                    "median_absolute_error",
+                    "accuracy",
+                    "roc_auc",
+                    "average_precision",
+                    "precision",
+                    "recall",
                 ]:
                     self.metric = metric
+                elif metric == "auc":
+                    self.metric = "roc_auc"
                 else:
                     raise ValueError("The input metric value is not valid.")
 
@@ -1340,7 +1406,7 @@ class GLMNetCVRegressor:
 
     def fit(self, X_train, y_train):
         """
-        Function to initialize a ElasticNet model using (X, y).
+        Function to initialize a LogitNet model using (X, y).
 
         Parameters
         ----------
@@ -1367,14 +1433,21 @@ class GLMNetCVRegressor:
 
         return None
 
-    def predict(self, X_test, lamb=None):
+    def predict_proba(self, X_test, y_test=None, lamb=None):
         """
-        Function to return the prediction values
+        Function to return the prediction probabilities for positive classes.
+        Please note that it only reports the probability of the positive class,
+        while the sklearn one returns for both and slicing like pred_proba[:, 1]
+        is needed for positive class predictions. Note that y_test is optional while
+        it might not be available in validiation.
 
         Parameters
         ----------
         X_test: numpy.array or pandas.DataFrame
             Validation features data
+
+        y_test: numpy.array[int] or list[int], optional (default=None)
+            List of validation ground truth binary values [0, 1]
 
         lamb: array, optional (default=None)
         Values with shape (n_lambda,) of lambda from lambda_path_
@@ -1384,7 +1457,42 @@ class GLMNetCVRegressor:
         values greater than max(lambda_path_) or less than  min(lambda_path_)
         will be clipped.
         """
-        self.y_pred_ = self.model_.predict(X_test, lamb=lamb)
+        self.X_test_, self.y_test_ = self._dtest(X_test, y_test)
+        if self.sparse_matrix:
+            self.y_pred_proba_ = self.model_.predict_proba(df_to_csr(self.X_test_), lamb=lamb)[:, 1]
+        else:
+            self.y_pred_proba_ = self.model_.predict_proba(self.X_test_, lamb=lamb)[:, 1]
+
+        return self.y_pred_proba_
+
+    def predict(self, X_test, y_test=None, threshold=0.5, lamb=None):
+        """
+        Function to return the prediction classes based on the passed threshold.
+        The default threshold is set at 0.5 while you can find the optimum thresholds
+        based on different methods using BinaryClassificationMetrics.
+
+        Parameters
+        ----------
+        X_test: numpy.array or pandas.DataFrame
+            Validation features data
+
+        y_test: numpy.array[int] or list[int], optional (default=None)
+            List of validation ground truth binary values [0, 1]
+
+        threshold: float, optional (default=0.5)
+            Threshold to define classes based on probabilities.
+            predict_proba >= threshold would be defined as 1 else 0.
+
+        lamb: array, optional (default=None)
+        Values with shape (n_lambda,) of lambda from lambda_path_
+        from which to make predictions. If no values are provided (None),
+        the returned predictions will be those corresponding to lambda_best_.
+        The values of lamb must also be in the range of lambda_path_,
+        values greater than max(lambda_path_) or less than  min(lambda_path_)
+        will be clipped.
+        """
+        y_pred_proba = self.predict_proba(X_test, y_test=y_test, lamb=lamb)
+        self.y_pred_ = [1 if p >= threshold else 0 for p in y_pred_proba]
 
         return self.y_pred_
 
@@ -1439,7 +1547,7 @@ class GLMNetCVRegressor:
         title=None,
         save_path=None,
     ):
-        """Function to plot GLMNetCVRegressor cross-validation results.
+        """Function to plot GLMNetCVClassfier cross-validation results.
 
         Parameters
         ----------
@@ -1525,7 +1633,7 @@ class GLMNetCVRegressor:
         yscale=None,
         save_path=None,
     ):
-        """Function to plot GLMNetCVRegressor coefficients' paths.
+        """Function to plot GLMNetCVClassfier coefficients' paths.
 
         Parameters
         ----------
@@ -1599,8 +1707,8 @@ class GLMNetCVRegressor:
         X_train: numpy.array or Pandas DataFrame
             Training features data
 
-        y_train: numpy.array[float] or list[float]
-            List of training ground truth values
+        y_train: numpy.array[int] or list[int]
+            List of training ground truth binary values [0, 1]
         """
         if isinstance(X_train, np.ndarray):
             self.X_train = pd.DataFrame(
@@ -1631,8 +1739,8 @@ class GLMNetCVRegressor:
         X_test: numpy.array or Pandas DataFrame
             Testing features data
 
-        y_test: numpy.array[float] or list[float]
-            List of testing ground truth values
+        y_test: numpy.array[int] or list[int]
+            List of testing ground truth binary values [0, 1]
         """
         if isinstance(X_test, np.ndarray):
             self.X_test = pd.DataFrame(X_test, columns=[f"F_{i}" for i in range(X_test.shape[1])])
@@ -1652,10 +1760,10 @@ class GLMNetCVRegressor:
 
     def _model(self):
         """
-        Function to initialize a ElasticNet model.
+        Function to initialize a LogitNet model.
         """
 
-        model = glmnet.ElasticNet(
+        model = glmnet.LogitNet(
             alpha=self.alpha,
             n_lambda=self.n_lambda,
             min_lambda_ratio=self.min_lambda_ratio,
@@ -1696,6 +1804,7 @@ class GLMNetCVRegressor:
             zip(
                 [self.X_train_.columns.tolist()[i] for i in idx],
                 [self.model_.coef_.reshape(-1, self.model_.coef_.shape[-1])[0][i] for i in idx],
+                #                 [self.model_.coef_[0][i] for i in idx],
             ),
         )
